Fixes various minor isues with image loading and has minor changes for the marketitem layout.
Fixes various mainor isues with image loading. Image loading now uses a better regex for checking whether or not a key is a folder or an actual data item. Images will now not try and load from the handler of the marketadapter of they are not in view (causing images to be loaded in recycled imageviews). The marketitem layout now has the title text centered and bolded.